+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [8.4.0]
 ### Added
+- Update metadata aggregator to add description to image variables from image processor and test tig forge processor
 ### Deprecated
 ### Removed
 ### Fixed